refactor: 回答詳細周辺のページ責務を分離する
回答詳細と周辺ページでも page、PageContent、View の境界を揃え、client side の SWR と表示 JSX が page.tsx に混ざらないようにする。
