+++ src/app/(protected)/(standard)/forms/[formId]/answers/[answerId]/_components/AnswerDetailsPageContent.tsx
@@ -1,0 +1,87 @@
+'use client';
+
+import {
+  getOptionalQueryData,
+  getRequiredQueryGroupError,
+  isQueryGroupReady,
+} from '@/app/_swr/queryState';
+import { useApiQuery } from '@/app/_swr/useApiQuery';
+import ErrorDialog from '@/app/_components/ErrorDialog';
+import LoadingCircular from '@/app/_components/LoadingCircular';
+import AnswerDetailsPageView from './AnswerDetailsPageView';
+import type { AnswerDetailsPageData } from './AnswerDetailsPageView';
+
+const AnswerDetailsPageContent = ({
+  formId,
+  answerId,
+}: {
+  formId: string;
+  answerId: string;
+}) => {
+  const answerQuery = useApiQuery(
+    '/api/v1/forms/{form_id}/answers/{answer_id}',
+    {
+      path: { form_id: formId, answer_id: answerId },
+    },
+    { refreshInterval: 1000 }
+  );
+
+  const { data: answer } = answerQuery;
+
+  const formQuery = useApiQuery(
+    '/api/v1/forms/{id}',
+    {
+      path: { id: answer?.form_id ?? '' },
+    },
+    { refreshInterval: 1000 }
+  );
+
+  const currentUserQuery = useApiQuery('/api/v1/users/me');
+
+  const messagesQuery = useApiQuery(
+    '/api/v1/forms/{form_id}/answers/{answer_id}/messages',
+    {
+      path: { form_id: formId, answer_id: answerId },
+    },
+    { refreshInterval: 1000 }
+  );
+
+  const commentsQuery = useApiQuery(
+    '/api/v1/forms/{form_id}/answers/{answer_id}/comments',
+    {
+      path: { form_id: formId, answer_id: answerId },
+    },
+    { refreshInterval: 1000 }
+  );
+
+  const requiredQueries = {
+    answer: answerQuery,
+    form: formQuery,
+    messages: messagesQuery,
+    comments: commentsQuery,
+  };
+  const queryError = getRequiredQueryGroupError(requiredQueries);
+
+  if (queryError !== undefined) {
+    return <ErrorDialog />;
+  }
+
+  if (!isQueryGroupReady(requiredQueries)) {
+    return <LoadingCircular />;
+  }
+
+  const currentUser = getOptionalQueryData(currentUserQuery);
+  const data: AnswerDetailsPageData = {
+    answer: requiredQueries.answer.data,
+    form: requiredQueries.form.data,
+    messages: requiredQueries.messages.data,
+    comments: requiredQueries.comments.data,
+    currentUserId: currentUser?.id,
+  };
+
+  return (
+    <AnswerDetailsPageView formId={formId} answerId={answerId} data={data} />
+  );
+};
+
+export default AnswerDetailsPageContent;
