+++ src/app/(protected)/(standard)/users/[userId]/_components/UserPageContent.tsx
@@ -1,0 +1,23 @@
+import UserView from './UserView';
+import type {
+  GetUserNotificationSettingsResponse,
+  GetUsersResponse,
+} from '@/lib/api-types';
+
+const UserPageContent = ({
+  user,
+  userId,
+  notificationSettings,
+}: {
+  user: GetUsersResponse;
+  userId: string;
+  notificationSettings: GetUserNotificationSettingsResponse;
+}) => (
+  <UserView
+    user={user}
+    userId={userId}
+    notificationSettings={notificationSettings}
+  />
+);
+
+export default UserPageContent;
