+++ src/app/(protected)/admin/answer/[answerId]/page.tsx
@@ -1,181 +1,14 @@
-'use client';
+import AdminAnswerPageContent from './_components/AdminAnswerPageContent';
+import type { Metadata } from 'next';
 
-import { Stack } from '@mui/material';
-import { use } from 'react';
-import {
-  getRequiredQueryGroupError,
-  isQueryGroupReady,
-} from '@/app/_swr/queryState';
-import { useApiQuery } from '@/app/_swr/useApiQuery';
-import ErrorDialog from '@/app/_components/ErrorDialog';
-import LoadingCircular from '@/app/_components/LoadingCircular';
-import Messages from '@/app/(protected)/_components/Messages';
-import StandardAnswerDetails from '@/app/(protected)/(standard)/forms/[formId]/answers/[answerId]/_components/AnswerDetails';
-import StandardAnswerMeta from '@/app/(protected)/(standard)/forms/[formId]/answers/[answerId]/_components/AnswerMeta';
-import Comments from './_components/Comments';
-import {
-  AdminAnswerLabelManagementButton,
-  AdminAnswerLabels,
-  AdminAnswerTitle,
-} from './_components/AnswerDetails';
-import { usePageTitle } from '@/hooks/usePageTitle';
-import type {
-  AnswerComment,
-  GetAnswerLabelsResponse,
-  GetAnswersResponse,
-  GetFormResponse,
-  GetMessagesResponse,
-} from '@/lib/api-types';
-
-type AdminAnswer = GetAnswersResponse[number];
-
-type AdminAnswerPageData = {
-  answer: AdminAnswer;
-  form: GetFormResponse;
-  labels: GetAnswerLabelsResponse;
-  messages: GetMessagesResponse;
-  comments: AnswerComment[];
+export const metadata: Metadata = {
+  title: '回答管理 | Seichi Portal',
 };
 
-const AdminAnswerPageView = ({
-  answerId,
-  data,
-}: {
-  answerId: string;
-  data: AdminAnswerPageData;
-}) => (
-  <Stack
-    direction="column"
-    spacing={4}
-    sx={{
-      width: '100%',
-      justifyContent: 'flex-start',
-      alignItems: 'stretch',
-    }}
-  >
-    <AdminAnswerTitle answer={data.answer} />
-    <StandardAnswerMeta
-      answer={data.answer}
-      labelsSlot={
-        <AdminAnswerLabels labelOptions={data.labels} answer={data.answer} />
-      }
-      extraActions={<AdminAnswerLabelManagementButton />}
-      messageAction={
-        <Messages
-          messages={data.messages}
-          formId={data.answer.form_id}
-          answerId={answerId}
-          title="メッセージ"
-          triggerLabel={`回答者にメッセージを送信 (${data.messages.length})`}
-        />
-      }
-    />
-    <StandardAnswerDetails
-      answer={data.answer}
-      questions={data.form.questions}
-    />
-    <Comments
-      comments={data.comments}
-      formId={data.answer.form_id}
-      answerId={answerId}
-      currentUserId={undefined}
-      showDeleteButton
-    />
-  </Stack>
-);
+const Home = async ({ params }: { params: Promise<{ answerId: string }> }) => {
+  const { answerId } = await params;
 
-const Home = ({ params }: { params: Promise<{ answerId: string }> }) => {
-  usePageTitle('回答管理');
-  const { answerId } = use(params);
-  const allAnswersQuery = useApiQuery('/api/v1/forms/answers', undefined, {
-    refreshInterval: 1000,
-  });
-
-  const { data: allAnswers } = allAnswersQuery;
-  const answers = allAnswers?.find((a) => a.id === answerId);
-
-  const formQuery = useApiQuery(
-    '/api/v1/forms/{id}',
-    {
-      path: { id: answers?.form_id ?? '' },
-    },
-    { refreshInterval: 1000 }
-  );
-
-  const labelsQuery = useApiQuery('/api/v1/labels/answers');
-
-  const messagesQuery = useApiQuery(
-    '/api/v1/forms/{form_id}/answers/{answer_id}/messages',
-    {
-      path: {
-        form_id: answers?.form_id ?? '',
-        answer_id: answerId,
-      },
-    },
-    { refreshInterval: 1000 }
-  );
-
-  const commentsQuery = useApiQuery(
-    '/api/v1/forms/{form_id}/answers/{answer_id}/comments',
-    {
-      path: {
-        form_id: answers?.form_id ?? '',
-        answer_id: answerId,
-      },
-    },
-    { refreshInterval: 1000 }
-  );
-
-  const answerListQueries = { allAnswers: allAnswersQuery };
-  const answerListError = getRequiredQueryGroupError(answerListQueries);
-
-  if (answerListError !== undefined) {
-    return <ErrorDialog />;
-  }
-
-  if (!isQueryGroupReady(answerListQueries)) {
-    return <LoadingCircular />;
-  }
-
-  const answer = answerListQueries.allAnswers.data.find(
-    (a) => a.id === answerId
-  );
-
-  if (answer === undefined) {
-    return (
-      <ErrorDialog
-        status={404}
-        title="回答が見つかりません"
-        message="指定された回答は存在しないか、表示できません。"
-      />
-    );
-  }
-
-  const detailQueries = {
-    form: formQuery,
-    labels: labelsQuery,
-    messages: messagesQuery,
-    comments: commentsQuery,
-  };
-  const detailError = getRequiredQueryGroupError(detailQueries);
-
-  if (detailError !== undefined) {
-    return <ErrorDialog />;
-  }
-
-  if (!isQueryGroupReady(detailQueries)) {
-    return <LoadingCircular />;
-  }
-
-  const data: AdminAnswerPageData = {
-    answer,
-    form: detailQueries.form.data,
-    labels: detailQueries.labels.data,
-    messages: detailQueries.messages.data,
-    comments: detailQueries.comments.data,
-  };
-
-  return <AdminAnswerPageView answerId={answerId} data={data} />;
+  return <AdminAnswerPageContent answerId={answerId} />;
 };
 
 export default Home;
